pyproject.toml
[project]
name = "python-blockchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "cryptography>=44.0.3",
    "fastapi[all]>=0.115.12",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pycryptodome>=3.22.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.8",
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
lint.select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "B",     # flake8-bugbear
    "I",     # isort
    "C4",    # flake8-comprehensions
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "T20",   # flake8-print
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_incomplete_defs = true
no_implicit_optional = true
